requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
redis>=5.0
djangorestframework-gis>=0.19
gunicorn>=21.0
whitenoise>=6.0
